Separate gammasimp from combsimp
Separated gamma function simplification(gammasimp) from combsimp. Also
separated _gamma_as_comb from combsimp and refactored its algorithm,
adding detection of inverse of binomial coefficient. See
sympy/sympy#9785 and sympy/sympy#13132.
