+++ sympy/concrete/tests/test_gosper.py
@@ -43,9 +43,9 @@ def test_gosper_sum():
     # issue 6033:
     assert gosper_sum(
         n*(n + a + b)*a**n*b**n/(factorial(n + a)*factorial(n + b)), \
-        (n, 0, m)) == -a*b*(exp(m*log(a))*exp(m*log(b))*factorial(a)* \
-        factorial(b) - factorial(a + m)*factorial(b + m))/(factorial(a)* \
-        factorial(b)*factorial(a + m)*factorial(b + m))
+        (n, 0, m)).simplify() == -exp(m*log(a) + m*log(b))*gamma(a + 1) \
+        *gamma(b + 1)/(gamma(a)*gamma(b)*gamma(a + m + 1)*gamma(b + m + 1)) \
+        + 1/(gamma(a)*gamma(b))
 
 
 def test_gosper_sum_indefinite():
