+++ sympy/functions/combinatorial/factorials.py
@@ -19,7 +19,9 @@ class CombinatorialFunction(Function):
     """Base class for combinatorial functions. """
 
     def _eval_simplify(self, ratio, measure):
-        from sympy.simplify.simplify import combsimp
+        from sympy.simplify.combsimp import combsimp
+        # combinatorial function with non-integer arguments is
+        # automatically passed to gammasimp
         expr = combsimp(self)
         if measure(expr) <= ratio*measure(self):
             return expr
