+++ sympy/integrals/tests/test_meijerint.py
@@ -321,7 +321,7 @@ def test_lookup_table():
 
 def test_branch_bug():
     from sympy import powdenest, lowergamma
-    # TODO combsimp cannot prove that the factor is unity
+    # TODO gammasimp cannot prove that the factor is unity
     assert powdenest(integrate(erf(x**3), x, meijerg=True).diff(x),
            polar=True) == 2*erf(x**3)*gamma(S(2)/3)/3/gamma(S(5)/3)
     assert integrate(erf(x**3), x, meijerg=True) == \
@@ -339,7 +339,7 @@ def test_linear_subs():
 def test_probability():
     # various integrals from probability theory
     from sympy.abc import x, y
-    from sympy import symbols, Symbol, Abs, expand_mul, combsimp, powsimp, sin
+    from sympy import symbols, Symbol, Abs, expand_mul, gammasimp, powsimp, sin
     mu1, mu2 = symbols('mu1 mu2', real=True, nonzero=True, finite=True)
     sigma1, sigma2 = symbols('sigma1 sigma2', real=True, nonzero=True,
                              finite=True, positive=True)
@@ -408,10 +408,10 @@ def test_probability():
         /gamma(alpha)/gamma(beta)
     assert integrate(betadist, (x, 0, oo), meijerg=True) == 1
     i = integrate(x*betadist, (x, 0, oo), meijerg=True, conds='separate')
-    assert (combsimp(i[0]), i[1]) == (alpha/(beta - 1), 1 < beta)
+    assert (gammasimp(i[0]), i[1]) == (alpha/(beta - 1), 1 < beta)
     j = integrate(x**2*betadist, (x, 0, oo), meijerg=True, conds='separate')
     assert j[1] == (1 < beta - 1)
-    assert combsimp(j[0] - i[0]**2) == (alpha + beta - 1)*alpha \
+    assert gammasimp(j[0] - i[0]**2) == (alpha + beta - 1)*alpha \
         /(beta - 2)/(beta - 1)**2
 
     # Beta distribution
@@ -441,7 +441,7 @@ def test_probability():
     assert simplify(integrate(x*chisquared, (x, 0, oo), meijerg=True)) == k
     assert simplify(integrate(x**2*chisquared, (x, 0, oo), meijerg=True)) == \
         k*(k + 2)
-    assert combsimp(integrate(((x - k)/sqrt(2*k))**3*chisquared, (x, 0, oo),
+    assert gammasimp(integrate(((x - k)/sqrt(2*k))**3*chisquared, (x, 0, oo),
                     meijerg=True)) == 2*sqrt(2)/sqrt(k)
 
     # Dagum distribution
@@ -525,7 +525,7 @@ def test_probability():
 
     # misc tests
     k = Symbol('k', positive=True)
-    assert combsimp(expand_mul(integrate(log(x)*x**(k - 1)*exp(-x)/gamma(k),
+    assert gammasimp(expand_mul(integrate(log(x)*x**(k - 1)*exp(-x)/gamma(k),
                               (x, 0, oo)))) == polygamma(0, k)
