+++ sympy/simplify/tests/test_hyperexpand.py
@@ -346,7 +346,7 @@ def can_do_meijer(a1, a2, b1, b2, numeric=True):
 
 @slow
 def test_meijerg_expand():
-    from sympy import combsimp, simplify
+    from sympy import gammasimp, simplify
     # from mpmath docs
     assert hyperexpand(meijerg([[], []], [[0], []], -z)) == exp(z)
 
@@ -396,7 +396,7 @@ def test_meijerg_expand():
                   (meijerg([0, 2], [], [], [-1, 1], z), True))
 
     # Test that the simplest possible answer is returned:
-    assert combsimp(simplify(hyperexpand(
+    assert gammasimp(simplify(hyperexpand(
         meijerg([1], [1 - a], [-a/2, -a/2 + S(1)/2], [], 1/z)))) == \
         -2*sqrt(pi)*(sqrt(z + 1) + 1)**a/a
 
@@ -564,7 +564,7 @@ def test_meijerg_with_Floats():
 
 
 def test_lerchphi():
-    from sympy import combsimp, exp_polar, polylog, log, lerchphi
+    from sympy import gammasimp, exp_polar, polylog, log, lerchphi
     assert hyperexpand(hyper([1, a], [a + 1], z)/a) == lerchphi(z, 1, a)
     assert hyperexpand(
         hyper([1, a, a], [a + 1, a + 1], z)/a**2) == lerchphi(z, 2, a)
@@ -572,11 +572,11 @@ def test_lerchphi():
         lerchphi(z, 3, a)
     assert hyperexpand(hyper([1] + [a]*10, [a + 1]*10, z)/a**10) == \
         lerchphi(z, 10, a)
-    assert combsimp(hyperexpand(meijerg([0, 1 - a], [], [0],
+    assert gammasimp(hyperexpand(meijerg([0, 1 - a], [], [0],
         [-a], exp_polar(-I*pi)*z))) == lerchphi(z, 1, a)
-    assert combsimp(hyperexpand(meijerg([0, 1 - a, 1 - a], [], [0],
+    assert gammasimp(hyperexpand(meijerg([0, 1 - a, 1 - a], [], [0],
         [-a, -a], exp_polar(-I*pi)*z))) == lerchphi(z, 2, a)
-    assert combsimp(hyperexpand(meijerg([0, 1 - a, 1 - a, 1 - a], [], [0],
+    assert gammasimp(hyperexpand(meijerg([0, 1 - a, 1 - a, 1 - a], [], [0],
         [-a, -a, -a], exp_polar(-I*pi)*z))) == lerchphi(z, 3, a)
 
     assert hyperexpand(z*hyper([1, 1], [2], z)) == -log(1 + -z)
